Refactor lots of stuff into RegisteredPluginStepsRetriever.cs DynamicsValue/fake-xrm-easy#122
