+++ src/FakeXrmEasy.Plugins/Pipeline/PluginStepRegistrationManager.cs
@@ -13,6 +13,9 @@ using Microsoft.Xrm.Sdk;
 
 namespace FakeXrmEasy.Pipeline
 {
+    /// <summary>
+    /// Internal helper class that contains logic to register new plugin steps
+    /// </summary>
     internal class PluginStepRegistrationManager
     {
         internal static Guid RegisterPluginStepInternal<TPlugin>(IXrmFakedContext context,
