+++ tests/FakeXrmEasy.Plugins.Tests/IXrmFakedContextPipelineExtensions/GetStepsForStageWithRetrieveMultipleTests.cs
@@ -29,7 +29,7 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
         [Fact]
         public void Should_return_empty_list_of_steps_if_none_were_registered()
         {
-            var steps = _context.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_createRequest.RequestName, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous, _createRequest);
+            var steps = RegisteredPluginStepsRetriever.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_context, _createRequest.RequestName, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous, _createRequest);
             Assert.Empty(steps);
         }
 
@@ -49,7 +49,7 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
                 Mode = mode
             });
 
-            var steps = _context.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(requestName, stage, mode, _createRequest);
+            var steps = RegisteredPluginStepsRetriever.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_context, requestName, stage, mode, _createRequest);
             Assert.Single(steps);
 
             var pluginStep = steps.FirstOrDefault();
@@ -82,7 +82,7 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
                 Mode = mode
             });
 
-            var steps = _context.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(requestName, stage, mode, _createRequest);
+            var steps = RegisteredPluginStepsRetriever.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_context, requestName, stage, mode, _createRequest);
             Assert.Single(steps);
 
             var pluginStep = steps.FirstOrDefault();
@@ -114,7 +114,7 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
                 Mode = mode
             });
 
-            var steps = _context.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(requestName, stage, mode, _createRequest);
+            var steps = RegisteredPluginStepsRetriever.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_context, requestName, stage, mode, _createRequest);
             Assert.Empty(steps);
         }
 
@@ -143,7 +143,7 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
                 Rank = 1
             });
 
-            var steps = _context.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(requestName, stage, mode, _createRequest).ToList();
+            var steps = RegisteredPluginStepsRetriever.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_context, requestName, stage, mode, _createRequest).ToList();
             Assert.Equal(2, steps.Count);
 
             var firstPluginStep = steps[0];
@@ -184,7 +184,7 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
 
             _target.Name = "Some name";
 
-            var steps = _context.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(requestName, stage, mode, _createRequest);
+            var steps = RegisteredPluginStepsRetriever.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_context, requestName, stage, mode, _createRequest);
             Assert.Single(steps);
 
             var pluginStep = steps.FirstOrDefault();
@@ -220,7 +220,7 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
 
             _target.AccountCategoryCode = new OptionSetValue(0);
 
-            var steps = _context.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(requestName, stage, mode, _createRequest);
+            var steps = RegisteredPluginStepsRetriever.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_context, requestName, stage, mode, _createRequest);
             Assert.Single(steps);
 
             var pluginStep = steps.FirstOrDefault();
@@ -255,7 +255,7 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
                 FilteringAttributes = new string[] { "name" }
             });
 
-            var steps = _context.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(requestName, stage, mode, _createRequest);
+            var steps = RegisteredPluginStepsRetriever.GetPluginStepsForOrganizationRequestWithRetrieveMultiple(_context, requestName, stage, mode, _createRequest);
             Assert.Empty(steps);
         }
